src/python/requirements.txt
torchvision
numpy
scipy
tensorboard
tensorboard_logger
sklearn
scipy
opencv-python
matplotlib
nltk
cffi